Move WeaponConfig into weapon cpp/h
