+++ source/ConfigFiles.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <G3D/G3D.h>
+#include "Weapon.h"
 #include "TargetEntity.h"
 
 template <class T>
@@ -14,173 +15,6 @@ template <class T>
 static bool operator==(Array<T> a1, Array<T> a2) {
 	return !(a1 != a2);
 }
-
-
-
-
-
-/** Weapon configuration class */
-class WeaponConfig {
-public:
-	String	id = "default";												///< Id by which to refer to this weapon
-	int		maxAmmo = 10000;											///< Max ammo (clicks) allowed per trial (set large for laser mode)
-	float	firePeriod = 0.5;											///< Minimum fire period (set to 0 for laser mode)
-	bool	autoFire = false;											///< Fire repeatedly when mouse is held? (set true for laser mode)
-	float	damagePerSecond = 2.0f;										///< Damage per second delivered (compute shot damage as damagePerSecond/firePeriod)
-	String	fireSound = "sound/fpsci_fire_100ms.wav"; 					///< Sound to play on fire
-	float	fireSoundVol = 1.0f;										///< Volume for fire sound
-	bool	renderModel = false;										///< Render a model for the weapon?
-	bool	hitScan = true;												///< Is the weapon a projectile or hitscan
-
-	//Vector3	muzzleOffset = Vector3(0, 0, 0);							///< Offset to the muzzle of the weapon model
-	//bool	renderMuzzleFlash = false;									///< Render a muzzle flash when the weapon fires?
-
-	bool	renderBullets = false;										///< Render bullets leaving the weapon
-	float	bulletSpeed = 100.0f;										///< Speed to draw at for rendered rounds (in m/s)
-	float	bulletGravity = 0.0f;										///< Gravity to use for bullets (default is no droop)
-	Vector3 bulletScale = Vector3(0.05f, 0.05f, 2.f);					///< Scale to use on bullet object
-	Color3  bulletColor = Color3(5, 4, 0);								///< Color/power for bullet emissive texture
-	Vector3 bulletOffset = Vector3(0,0,0);								///< Offset to start the bullet from (along the look direction)
-	
-	bool	renderDecals = true;										///< Render decals when the shots miss?
-	String	missDecal = "bullet-decal-256x256.png";						///< The decal to place where the shot misses
-	String	hitDecal = "";												///< The decal to place where the shot hits
-	int		missDecalCount = 2;											///< Number of miss decals to draw
-	float	missDecalScale = 1.0f;										///< Scale to apply to the miss decal
-	float	hitDecalScale = 1.0f;										///< Scale to apply to the hit decal
-	float	hitDecalDurationS = 0.1f;									///< Duration to show the hit decal for (in seconds)
-	float	hitDecalColorMult = 2.0f;									///< "Encoding" field (aka color multiplier) for hit decal
-	
-	float	fireSpreadDegrees = 0;										///< The spread of the fire
-	String  fireSpreadShape = "uniform";								///< The shape of the fire spread distribution
-	float	damageRollOffAim = 0;										///< Damage roll off w/ aim
-	float	damageRollOffDistance = 0;									///< Damage roll of w/ distance
-
-	float	scopeFoV = 0.0f;											///< Field of view when scoped
-	bool	scopeToggle = false;										///< Scope toggle behavior
-	//String reticleImage;												///< Reticle image to show for this weapon
-
-	float	kickAngleDegrees = 0.0f;									///< Angle for the weapon to kick when fired
-	float	kickDuration = 0.0f;										///< Kick duration
-
-	ArticulatedModel::Specification modelSpec;							///< Model to use for the weapon (must be specified when renderModel=true)
-
-	/** Returns true if firePeriod == 0 and autoFire == true */
-	bool isContinuous() const {
-		return firePeriod == 0 && autoFire;
-	}
-
-	WeaponConfig() {}
-
-	/** Load from Any */
-	WeaponConfig(const Any& any) {
-		int settingsVersion = 1;
-		AnyTableReader reader(any);
-		reader.getIfPresent("settingsVersion", settingsVersion);
-
-		switch (settingsVersion) {
-		case 1:
-			reader.getIfPresent("id", id);
-
-			reader.getIfPresent("maxAmmo", maxAmmo);
-			reader.getIfPresent("firePeriod", firePeriod);
-			reader.getIfPresent("autoFire", autoFire);
-			reader.getIfPresent("damagePerSecond", damagePerSecond);
-			reader.getIfPresent("fireSound", fireSound);
-			reader.getIfPresent("fireSoundVol", fireSoundVol);
-			reader.getIfPresent("hitScan", hitScan);
-
-			reader.getIfPresent("renderModel", renderModel);			
-			if (renderModel) {
-				reader.get("modelSpec", modelSpec, "If \"renderModel\" is set to true within a weapon config then a \"modelSpec\" must be provided!");
-			}
-			else {
-				reader.getIfPresent("modelSpec", modelSpec);
-			}
-			
-			//reader.getIfPresent("muzzleOffset", muzzleOffset);
-			//reader.getIfPresent("renderMuzzleFlash", renderMuzzleFlash);
-
-			reader.getIfPresent("renderBullets", renderBullets);
-			reader.getIfPresent("bulletSpeed", bulletSpeed);
-			reader.getIfPresent("bulletGravity", bulletGravity);
-			reader.getIfPresent("bulletScale", bulletScale);
-			reader.getIfPresent("bulletColor", bulletColor);
-			reader.getIfPresent("bulletOffset", bulletOffset);
-
-			reader.getIfPresent("renderDecals", renderDecals);
-			reader.getIfPresent("missDecal", missDecal);
-			reader.getIfPresent("hitDecal", hitDecal);
-			reader.getIfPresent("missDecalCount", missDecalCount);
-			reader.getIfPresent("missDecalScale", missDecalScale);
-			reader.getIfPresent("hitDecalScale", hitDecalScale);
-			reader.getIfPresent("hitDecalDuration", hitDecalDurationS);
-			reader.getIfPresent("hitDecalColorMult", hitDecalColorMult);
-
-			reader.getIfPresent("fireSpreadDegrees", fireSpreadDegrees);
-			reader.getIfPresent("fireSpreadShape", fireSpreadShape);
-
-			reader.getIfPresent("damageRollOffAim", damageRollOffAim);
-			reader.getIfPresent("damageRollOffDistance", damageRollOffDistance);
-
-			reader.getIfPresent("scopeFoV", scopeFoV);
-			reader.getIfPresent("scopeToggle", scopeToggle);
-			reader.getIfPresent("kickAngleDegrees", kickAngleDegrees);
-			reader.getIfPresent("kickDuration", kickDuration);
-
-			//reader.getIfPresent("recticleImage", reticleImage);
-		default:
-			debugPrintf("Settings version '%d' not recognized in TargetConfig.\n", settingsVersion);
-			break;
-		}
-	}
-
-	Any toAny(const bool forceAll = false) const {
-		Any a(Any::TABLE);
-		WeaponConfig def;
-		a["id"] = id;
-		if (forceAll || def.maxAmmo != maxAmmo)								a["maxAmmo"] = maxAmmo;
-		if (forceAll || def.firePeriod != firePeriod)						a["firePeriod"] = firePeriod;
-		if (forceAll || def.autoFire != autoFire)							a["autoFire"] = autoFire;
-		if (forceAll || def.damagePerSecond != damagePerSecond)				a["damagePerSecond"] = damagePerSecond;
-		if (forceAll || def.fireSound != fireSound)							a["fireSound"] = fireSound;
-		if (forceAll || def.fireSoundVol != fireSoundVol)					a["fireSoundVol"] = fireSoundVol;
-		if (forceAll || def.hitScan != hitScan)								a["hitScan"] = hitScan;
-
-		if (forceAll || def.renderModel != renderModel)						a["renderModel"] = renderModel;
-		if (forceAll || !(def.modelSpec == modelSpec))						a["modelSpec"] = modelSpec;
-
-		//if (forceAll || def.muzzleOffset != muzzleOffset)					a["muzzleOffset"] = muzzleOffset;
-		//if (forceAll || def.renderMuzzleFlash != renderMuzzleFlash)			a["renderMuzzleFlash"] = renderMuzzleFlash;
-
-		if (forceAll || def.renderBullets != renderBullets)					a["renderBullets"] = renderBullets;
-		if (forceAll || def.bulletSpeed != bulletSpeed)						a["bulletSpeed"] = bulletSpeed;
-		if (forceAll || def.bulletGravity != bulletGravity)					a["bulletGravity"] = bulletGravity;
-		if (forceAll || def.bulletScale != bulletScale)						a["bulletScale"] = bulletScale;
-		if (forceAll || def.bulletColor != bulletColor)						a["bulletColor"] = bulletColor;
-		if (forceAll || def.bulletOffset != bulletOffset)					a["bulletOffset"] = bulletOffset;
-
-		if (forceAll || def.renderDecals != renderDecals)					a["renderDecals"] = renderDecals;
-		if (forceAll || def.missDecal != missDecal)							a["missDecal"] = missDecal;
-		if (forceAll || def.hitDecal != hitDecal)							a["hitDecal"] = hitDecal;
-		if (forceAll || def.missDecalCount != missDecalCount)				a["missDecalCount"] = missDecalCount;
-		if (forceAll || def.missDecalScale != missDecalScale)				a["missDecalScale"] = missDecalScale;
-		if (forceAll || def.hitDecalScale != hitDecalScale)					a["hitDecalScale"] = hitDecalScale;
-		if (forceAll || def.hitDecalDurationS != hitDecalDurationS)			a["hitDecalDuration"] = hitDecalDurationS;
-		if (forceAll || def.hitDecalColorMult != hitDecalColorMult)			a["hitDecalColorMult"] = hitDecalColorMult;
-
-		if (forceAll || def.fireSpreadDegrees != fireSpreadDegrees)			a["fireSpreadDegrees"] = fireSpreadDegrees;
-		if (forceAll || def.fireSpreadShape != fireSpreadShape)				a["fireSpreadShape"] = fireSpreadShape;
-		if (forceAll || def.damageRollOffAim != damageRollOffAim)			a["damageRollOffAim"] = damageRollOffAim;
-		if (forceAll || def.damageRollOffDistance != damageRollOffDistance)	a["damageRollOffDistance"] = damageRollOffDistance;
-		if (forceAll || def.scopeFoV != scopeFoV)							a["scopeFoV"] = scopeFoV;
-		if (forceAll || def.scopeToggle != scopeToggle)						a["scopeToggle"] = scopeToggle;
-		if (forceAll || def.kickAngleDegrees != kickAngleDegrees)			a["kickAngleDegrees"] = kickAngleDegrees;
-		if (forceAll || def.kickDuration != kickDuration)					a["kickDuration"] = kickDuration;
-
-		return a;
-	}
-};
 
 /** Class for representing a given target configuration */
 class TargetConfig : public ReferenceCountedObject {
